chore: split dialog into separate files
